Reuse switch
