+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/settings/DiagnosticsView.kt
@@ -11,23 +11,19 @@ import androidx.compose.material3.ListItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.algoritmico.passepartout.business.extensions.default
-import com.algoritmico.passepartout.business.extensions.runCatchingNonFatal
 import com.algoritmico.passepartout.context.isBetaSuggestedByAndroidAPI
 import com.algoritmico.passepartout.models.AppPreferenceKey
 import com.algoritmico.passepartout.models.AppPreferences
-import com.algoritmico.passepartout.observables.LocalErrorHandler
 import com.algoritmico.passepartout.observables.LocalUserPreferencesObservable
 import com.algoritmico.passepartout.observables.UserPreferencesObservable
 import com.algoritmico.passepartout.ui.theme.ThemeListSection
 import com.algoritmico.passepartout.ui.theme.ThemeNavigatingButton
 import com.algoritmico.passepartout.ui.theme.ThemeSwitchRow
-import kotlinx.coroutines.launch
 
 @Composable
 fun DiagnosticsView(
@@ -73,8 +69,9 @@ fun DiagnosticsView(
         }
         item {
             ThemeListSection(header = "Preferences") {
-                LogsPrivateDataRow(
-                    isChecked = preferences.logsPrivateData,
+                ThemeSwitchRow(
+                    title = "Include private data",
+                    checked = preferences.logsPrivateData,
                     onCheckedChange = { isChecked ->
                         userPreferencesObservable.updateLogsPrivateData(isChecked)
                     }
@@ -87,31 +84,6 @@ fun DiagnosticsView(
             }
         }
     }
-}
-
-@Composable
-private fun LogsPrivateDataRow(
-    isChecked: Boolean,
-    onCheckedChange: suspend (Boolean) -> Unit
-) {
-    val coroutineScope = rememberCoroutineScope()
-    val errorHandler = LocalErrorHandler.current
-
-    fun update(isChecked: Boolean) {
-        coroutineScope.launch {
-            runCatchingNonFatal {
-                onCheckedChange(isChecked)
-            }.onFailure {
-                errorHandler.report(it)
-            }
-        }
-    }
-
-    ThemeSwitchRow(
-        title = "Include private data",
-        checked = isChecked,
-        onCheckedChange = ::update
-    )
 }
 
 private suspend fun UserPreferencesObservable.updateLogsPrivateData(
