+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/settings/PreferencesView.kt
@@ -7,39 +7,29 @@ package com.algoritmico.passepartout.ui.settings
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.algoritmico.passepartout.business.extensions.runCatchingNonFatal
-import com.algoritmico.passepartout.observables.LocalErrorHandler
 import com.algoritmico.passepartout.observables.LocalUserPreferencesObservable
 import com.algoritmico.passepartout.ui.theme.ThemeListSection
 import com.algoritmico.passepartout.ui.theme.ThemeSwitchRow
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.launch
 
 @Composable
 fun PreferencesView(
     modifier: Modifier = Modifier,
     onAdvanced: () -> Unit
 ) {
-    val coroutineScope = rememberCoroutineScope()
-    val errorHandler = LocalErrorHandler.current
     val userPreferencesObservable = LocalUserPreferencesObservable.current
     Column(
         modifier = modifier.fillMaxSize()
     ) {
         ThemeListSection {
-            PreferenceSwitchRow("DNS fallback", userPreferencesObservable.dnsFallback) {
-                coroutineScope.launch {
-                    runCatchingNonFatal {
-                        userPreferencesObservable.toggleDnsFallback()
-                    }.onFailure {
-                        errorHandler.report(it)
-                    }
+            ThemeSwitchRow(
+                title = "DNS fallback",
+                supportingText = "Fall back to CloudFlare servers when the VPN does not provide DNS settings.",
+                checkedFlow = userPreferencesObservable.dnsFallback,
+                onCheckedChange = {
+                    userPreferencesObservable.toggleDnsFallback()
                 }
-            }
+            )
             // Hide "Advanced" because there are no actionable config flags
 //            ThemeNavigatingButton(
 //                title = "Advanced",
@@ -47,21 +37,4 @@ fun PreferencesView(
 //            )
         }
     }
-}
-
-@Composable
-private fun PreferenceSwitchRow(
-    title: String,
-    checkedFlow: Flow<Boolean>,
-    onCheckedChange: (Boolean) -> Unit
-) {
-    val checked by checkedFlow.collectAsStateWithLifecycle(
-        initialValue = false
-    )
-    ThemeSwitchRow(
-        title = title,
-        supportingText = "Fall back to CloudFlare servers when the VPN does not provide DNS settings.",
-        checked = checked,
-        onCheckedChange = onCheckedChange
-    )
 }
