+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/theme/ThemeSwitchRow.kt
@@ -9,7 +9,37 @@ import androidx.compose.material3.ListItem
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.algoritmico.passepartout.business.extensions.runCatchingNonFatal
+import com.algoritmico.passepartout.observables.LocalErrorHandler
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.launch
+
+@Composable
+fun ThemeSwitchRow(
+    title: String,
+    checkedFlow: Flow<Boolean>,
+    modifier: Modifier = Modifier,
+    supportingText: String? = null,
+    enabled: Boolean = true,
+    initialValue: Boolean = false,
+    onCheckedChange: suspend (Boolean) -> Unit
+) {
+    val checked by checkedFlow.collectAsStateWithLifecycle(
+        initialValue = initialValue
+    )
+    ThemeSwitchRow(
+        title = title,
+        checked = checked,
+        modifier = modifier,
+        supportingText = supportingText,
+        enabled = enabled,
+        onCheckedChange = onCheckedChange
+    )
+}
 
 @Composable
 fun ThemeSwitchRow(
@@ -18,8 +48,21 @@ fun ThemeSwitchRow(
     modifier: Modifier = Modifier,
     supportingText: String? = null,
     enabled: Boolean = true,
-    onCheckedChange: (Boolean) -> Unit
+    onCheckedChange: suspend (Boolean) -> Unit
 ) {
+    val coroutineScope = rememberCoroutineScope()
+    val errorHandler = LocalErrorHandler.current
+
+    fun update(isChecked: Boolean) {
+        coroutineScope.launch {
+            runCatchingNonFatal {
+                onCheckedChange(isChecked)
+            }.onFailure {
+                errorHandler.report(it)
+            }
+        }
+    }
+
     ListItem(
         headlineContent = {
             Text(title)
@@ -33,13 +76,13 @@ fun ThemeSwitchRow(
             Switch(
                 checked = checked,
                 enabled = enabled,
-                onCheckedChange = onCheckedChange
+                onCheckedChange = ::update
             )
         },
         modifier = modifier.clickable(
             enabled = enabled
         ) {
-            onCheckedChange(!checked)
+            update(!checked)
         }
     )
 }
